Make it work on 32-bit architectures...

+++ src/telegram/api/common_commands.go
@@ -1,7 +1,7 @@
 package api
 
 type SendMessage struct {
-    ChatId                  int         `json:"chat_id"`
+    ChatId                  int64       `json:"chat_id"`
     Text                    string      `json:"text"`
     ParseMode               string      `json:"parse_mode,omitempty"`
     DisableWebPagePreview   *bool       `json:"disable_web_page_preview"`
